requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements-dev.txt requirements-dev.in
#
asgiref==3.5.2
    # via django
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements-dev.in
django==4.0.4
    # via
    #   -r requirements-dev.in
    #   django-extensions
    #   django-guardian
    #   django-import-export
    #   djangorestframework
django-admin-list-filter-dropdown==1.0.3
    # via -r requirements-dev.in
django-adminfilters==1.9.0
    # via -r requirements-dev.in
django-extensions==3.1.5
    # via -r requirements-dev.in
django-guardian==2.4.0
    # via -r requirements-dev.in
django-import-export==2.8.0
    # via -r requirements-dev.in
djangorestframework==3.13.1
    # via -r requirements-dev.in
environs==9.5.0
    # via -r requirements-dev.in
et-xmlfile==1.1.0
    # via openpyxl
markuppy==1.14
    # via tablib
marshmallow==3.15.0
    # via environs
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via marshmallow
pyparsing==3.0.9
    # via packaging
python-decouple==3.6
    # via -r requirements-dev.in
python-dotenv==0.20.0
    # via environs
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via tablib
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
psycopg2 
    # postgresql database

django-crispy-forms
Pillow    
Markdown==3.3.7